+++ logs/http_request_properties.go
@@ -14,7 +14,10 @@
 
 package logs
 
-import "net/http"
+import (
+	"net/http"
+	"regexp"
+)
 
 // HTTPRequestProperties is an entity which contains the properties of an HTTP request
 type HTTPRequestProperties struct {
@@ -22,10 +25,19 @@ type HTTPRequestProperties struct {
 	Path       string
 	RemoteAddr string
 	UserAgent  string
+	UseRegex   bool // If true, interprets all properties as regex. Otherwise matches on equality
 }
 
 // Match returns true if the provided http.Request matches the defined request properties
 func (h HTTPRequestProperties) Match(r *http.Request) bool {
+	if h.UseRegex {
+		return h.MatchRegex(r)
+	}
+	return h.MatchEquality(r)
+}
+
+// MatchEquality returns true if the provided http.Request matches the defined request properties by equality
+func (h HTTPRequestProperties) MatchEquality(r *http.Request) bool {
 	if h.Method != "" && h.Method != r.Method {
 		return false
 	}
@@ -45,6 +57,39 @@ func (h HTTPRequestProperties) Match(r *http.Request) bool {
 	return true
 }
 
+// MatchRegex returns true if the provided http.Request matches the defined request properties by regex
+func (h HTTPRequestProperties) MatchRegex(r *http.Request) bool {
+	if h.Method != "" {
+		matched, _ := regexp.MatchString(h.Method, r.Method)
+		if !matched {
+			return false
+		}
+	}
+
+	if h.Path != "" {
+		matched, _ := regexp.MatchString(h.Path, r.URL.Path)
+		if !matched {
+			return false
+		}
+	}
+
+	if h.RemoteAddr != "" {
+		matched, _ := regexp.MatchString(h.RemoteAddr, r.RemoteAddr)
+		if !matched {
+			return false
+		}
+	}
+
+	if h.UserAgent != "" {
+		matched, _ := regexp.MatchString(h.UserAgent, r.UserAgent())
+		if !matched {
+			return false
+		}
+	}
+
+	return true
+}
+
 // NewAwsHealthCheckHTTPRequestProperties creates an HTTPRequestProperties object for a standard AWS ELB health checker
 //
 //	Path: The path that the health checks are performed on. If empty, "/version" is used as the default value.
@@ -52,7 +97,7 @@ func NewAwsHealthCheckHTTPRequestProperties(path string) HTTPRequestProperties {
 	if path == "" {
 		path = "/version"
 	}
-	return HTTPRequestProperties{Method: "GET", Path: path, UserAgent: "ELB-HealthChecker/2.0"}
+	return HTTPRequestProperties{Method: "GET", Path: path, UserAgent: "ELB-HealthChecker/*", UseRegex: true}
 }
 
 // NewOpenShiftHealthCheckHTTPRequestProperties creates an HTTPRequestProperties object for a standard OpenShift health checker
@@ -62,7 +107,7 @@ func NewOpenShiftHealthCheckHTTPRequestProperties(path string) HTTPRequestProper
 	if path == "" {
 		path = "/version"
 	}
-	return HTTPRequestProperties{Method: "GET", Path: path, UserAgent: "kube-probe/1.22+"}
+	return HTTPRequestProperties{Method: "GET", Path: path, UserAgent: "kube-probe/*", UseRegex: true}
 }
 
 // NewStandardHealthCheckHTTPRequestProperties creates a list of HTTPRequestProperties objects for known standard health checkers