+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Regex in HTTPRequestProperties [#20](https://github.com/rokwire/logging-library-go/issues/20)
 
 ## [2.0.0] - 2022-11-16
 ### Added